backend/requirements.txt
APIFlask==0.12.0
apispec==5.1.1
appnope==0.1.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
executing==0.8.3
filelock==3.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.4.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.34
stack-data==0.2.0
stevedore==3.5.0
toml==0.10.2
traitlets==5.1.1
virtualenv==20.13.4
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
